+++ RELEASES.md
@@ -1,3 +1,8 @@
+2.8.0 2023-04-29
+-----------------
+  - feat: happn-3 changePassword in client
+  - feat: happn-3 resetPassword supported by auth provider
+
 2.7.1 2023-04-14
 -----------------
   - fix: happner-2 rest stringify ignores special characters