+++ packages/happn-cluster/RELEASES.md
@@ -354,3 +354,8 @@ non-release 2017-12-14
 -----------------
   - feat: token revocation and logout
   - fix: token revocation cluster fix
+
+12.4.0 2023-04-29
+-----------------
+  - feat: changePassword in client
+  - feat: resetPassword supported by auth provider