+++ packages/happner-client/RELEASES.md
@@ -418,3 +418,8 @@
 12.6.1 2023-04-13
 -----------------
 - fix: rest payload stringify
+
+12.7.0 2023-04-29
+-----------------
+  - feat: happn-3 changePassword in client
+  - feat: happn-3 resetPassword supported by auth provider